Sécurisation avec un try catch de la requête d'insertion d'image et de la requête de modification d'image.Création du dossier 'élémentsgraphiques'.Changement du chemins des images du header. Ajout du script de suppression d'images dans le script de suppresssion_image.php.Insertion du champ de nom de fichier dans le formulaire de suppression d'image.Sécurisation des champs de formulaires.

+++ Modeles/suppression_image.php
@@ -1,26 +1,40 @@
 <?php
 
-
-
-// Utilisation du fichier config pour récupérer les variables d'environnement:
 require_once __DIR__ . '../../config.php';
 require_once __DIR__ . '../../vendor/autoload.php';
 
 $id = htmlspecialchars($_POST['id'],ENT_QUOTES);
+$filename = htmlspecialchars($_POST['filename'],ENT_QUOTES);
+
 
 try{
-$pdo = new PDO(DB_DSN, DB_USER, DB_PASSWORD);
-$statement = $pdo->prepare("DELETE FROM galerie_images WHERE id = :id");
-    $statement->bindValue(':id', $id, PDO::PARAM_INT);
-    $statement->execute(); 
-    echo 'Suppression effectuée.';
+    // Suppression du fichier dans son dossier images
+    $image_path = __DIR__ . '/../images/' . $filename;
+    if (unlink($image_path) === true && file_exists($image_path) === false) {
+  // Suppression de l'image en base de données
+        $pdo = new PDO(DB_DSN, DB_USER, DB_PASSWORD);
+        $statement = $pdo->prepare("DELETE FROM galerie_images WHERE id = :id");
+        $statement->bindValue(':id', $id, PDO::PARAM_INT);
+        $statement->execute(); 
+
+        echo 'Suppression effectuée.';
+        echo '<div>';
+        echo '<a href="../../index.php?page=admin"><button class="button-reservation-style">Retour page administrateur</button></a>';
+        echo '</div>';
+        exit;
+    } else {
+        throw new Exception("Impossible de supprimer le fichier image.");
+    }
+} catch(PDOException $PDOException){
     echo '<div>';
+    echo 'Erreur lors de la suppression de l\'image : ' . $PDOException->getMessage();
     echo '<a href="../../index.php?page=admin"><button class="button-reservation-style">Retour page administrateur</button></a>';
     echo '</div>';
     exit;
-
-}catch(PDOException $PDOException){
+    
+} catch (Exception $e) {
     echo '<div>';
+    echo 'Erreur lors de la suppression de l\'image : ' . $e->getMessage();
     echo '<a href="../../index.php?page=admin"><button class="button-reservation-style">Retour page administrateur</button></a>';
     echo '</div>';
     exit;
